refactor: rename model to models
