+++ oxide_editor/src/models/content_drawer.rs
